chore: rename resend package
