+++ packages/email-resend/.eslintignore
