+++ packages/email-resend/.eslintrc.cjs
