+++ packages/email-resend/.prettierignore
