+++ packages/email-resend/.swcrc
