+++ packages/email-resend/.swcrc-build
