+++ packages/email-resend/LICENSE.md
