+++ packages/email-resend/README.md
@@ -5,7 +5,7 @@ This adapter allows you to send emails using the [Resend](https://resend.com) RE
 ## Installation
 
 ```sh
-pnpm add @payloadcms/email-resend-rest`
+pnpm add @payloadcms/email-resend`
 ```
 
 ## Usage
@@ -18,7 +18,7 @@ pnpm add @payloadcms/email-resend-rest`
 
 ```ts
 // payload.config.js
-import { resendAdapter } from '@payloadcms/email-resend-rest'
+import { resendAdapter } from '@payloadcms/email-resend'
 
 export default buildConfig({
   email: resendAdapter({
